Workers/tests/submissions/Microsoft/WorkerGlobalScope_ErrorEvent_message.htm

<!DOCTYPE html>
<html>
<head>
<title> WorkerGlobalScope ErrorEvent.message </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    setup({timeout:500});
    var ErrorMessage = "Error Message";
    var FileName = './support/ErrorEvent.js';
    
    var t = async_test("Test Description: ErrorEvent message attribute represents the error message.");
                  
    var worker = new Worker(FileName);

    worker.onmessage = t.step_func(function(evt)
    {
        var err = evt.data;
        assert_not_equals(err.message, undefined);
        assert_not_equals(err.message.indexOf('Error Message'), -1);
        t.done();        
    });
    
    worker.postMessage(ErrorMessage);
</script>
</body>
</html>
